(feat) Added configurable queue tab section in queue table
