+++ packages/esm-outpatient-app/src/queue-entry-table-components/actions-menu.scss
@@ -1,0 +1,3 @@
+.menuItem {
+  max-width: none;
+}
