+++ packages/esm-outpatient-app/src/queue-entry-table-components/status-icon.component.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { Group, InProgress } from '@carbon/react/icons';
+
+function StatusIcon({ status }) {
+  switch (status) {
+    case 'waiting':
+      return <InProgress size={16} />;
+    case 'in service':
+      return <Group size={16} />;
+    case 'finished service':
+      return <Group size={16} />;
+    default:
+      return null;
+  }
+}
+
+export default StatusIcon;
