+++ packages/esm-outpatient-app/translations/en.json
@@ -1,10 +1,15 @@
 {
   "actions": "Actions",
+  "activeVisits": "Active Visits",
+  "add": "Add",
+  "addEntry": "Add entry",
+  "addisitToQueueTooltip": "Add",
   "addNewQueue": "Add new queue",
   "addNewService": "Add new service",
   "addPatientToList": "Add patient to list",
   "addPatientToQueue": "Add patient to queue",
   "addQueue": "Add queue",
+  "addVisitToQueue": "Add Visit To Queue?",
   "advancedSearch": "Advanced search",
   "age": "Age",
   "alistOfClients": "A list of clients waiting for ",
@@ -31,6 +36,7 @@
   "clearQueue": "Clear queue",
   "clinicMetrics": "Clinic metrics",
   "configurePriorities": "Please configure priorities to continue.",
+  "configureServices": "Please configure services to continue.",
   "contactDetails": "Contact Details",
   "currentVisit": "Current visit",
   "date": "Date",
@@ -56,6 +62,7 @@
   "fields": "of the following fields",
   "filter": "Filter",
   "filters": "Filters",
+  "filterTable": "Filter table",
   "firstName": "First name",
   "firstNameSort": "First name (a-z)",
   "futureScheduledVisits": "{count, plural, one{{count} visit} other{{count} visits}} scheduled for dates in the future",
@@ -63,8 +70,9 @@
   "gender": "Gender",
   "heartRate": "Heart rate",
   "height": "Height",
-  "hoursAnd": "hours and ",
+  "idNumber": "ID Number",
   "indication": "Indication",
+  "InQueue": "In Queue",
   "lastEncounter": "Last encounter",
   "lastName": "Last name",
   "lastNameSort": "Last name (a-z)",
@@ -82,6 +90,7 @@
   "missingVisitType": "Missing visit type",
   "moreMetrics": "See more metrics",
   "name": "Name",
+  "noActiveVisitsForLocation": "There are no active visits to display for this location.",
   "noAppointmentsFound": "No appointments found",
   "noConceptsAvailable": "No concepts available",
   "noEncountersFound": "No encounters found",
@@ -100,6 +109,8 @@
   "noServicesConfigured": "No services configured",
   "noStatusAvailable": "No Status Available",
   "notes": "Notes",
+  "NotInQueue": "Not In Queue",
+  "noVisitsFound": "No matching visits found",
   "noVisitTypes": "There are no visit types to display for this patient",
   "noVitalsFound": "No vitals found",
   "oldest": "Oldest first",
@@ -126,6 +137,8 @@
   "quantity": "Quantity",
   "queueAddedSuccessfully": "Queue addeded successfully",
   "queueDescription": "Queue description",
+  "queueEntryAddedSuccessfully": "Queue Entry Added Successfully",
+  "queueEntryAddFailed": "Error adding queue entry status",
   "queueEntryError": "Error adding patient to the queue",
   "queueEntryRemoved": "Queue entry removed",
   "queueEntryRemovedSuccessfully": "",
@@ -188,6 +201,7 @@
   "useTodaysDate": "Use today's date",
   "view": "View",
   "visitLocation": "Visit Location",
+  "visitStartTime": "Visit Time",
   "visitType": "Visit Type",
   "vitals": "Vitals",
   "vitalsForm": "Vitals form",
